+++ test/hook-app.js
@@ -60,6 +60,15 @@ t.test('Hook app', async t => {
   });
 
   app.addContextHook('dispatch:before', async ctx => {
+    const {name, req, res} = ctx.backend;
+    const middleware = ctx.req.query.get('middleware');
+    if (name !== 'server' || middleware !== '1') return;
+    res.writeHead(200, ['X-URL', req.url]);
+    res.end('Hello Middleware!');
+    return true;
+  });
+
+  app.addContextHook('dispatch:before', async ctx => {
     if (ctx.isWebSocket === false) return;
 
     const third = ctx.req.query.get('third');
@@ -157,6 +166,11 @@ t.test('Hook app', async t => {
     (await ua.getOk('/')).statusIs(200).bodyIs('Hello Mojo!');
     (await ua.getOk('/?fourth=1')).statusIs(200).bodyIs('Fourth hook');
     (await ua.getOk('/static/mojo/favicon.ico?fourth=1')).statusIs(200).bodyIsnt('Fourth hook');
+  });
+
+  await t.test('Server request hook', async () => {
+    (await ua.getOk('/')).statusIs(200).bodyIs('Hello Mojo!');
+    (await ua.getOk('/?middleware=1')).statusIs(200).headerIs('X-URL', '/?middleware=1').bodyIs('Hello Middleware!');
   });
 
   await t.test('Send hooks', async () => {